requirements.txt
telethon
python-dotenv
sqlalchemy
psycopg2-binary
dbt-postgres
ultralytics
fastapi
uvicorn
dagster
dagster-webserver
